+++ modules/mod_regex/regex.c
@@ -2314,7 +2314,7 @@ typedef struct
    value.  Assumes the variable `fail_stack'.  Probably should only
    be called from within `PUSH_FAILURE_POINT'.  */
 #define PUSH_FAILURE_ITEM(item)                     \
-  fail_stack.stack[fail_stack.avail++] = (fail_stack_elt_t) item
+  fail_stack.stack[fail_stack.avail++] = (fail_stack_elt_t) (size_t) item
 
 /* The complement operation.  Assumes `fail_stack' is nonempty.  */
 #define POP_FAILURE_ITEM() fail_stack.stack[--fail_stack.avail]
@@ -2481,10 +2481,10 @@ typedef struct
   DEBUG_PRINT_COMPILED_PATTERN (bufp, pat, pend);           \
                                     \
   /* Restore register info.  */                     \
-  high_reg = (unsigned) POP_FAILURE_ITEM ();                \
+  high_reg = (unsigned) (size_t) POP_FAILURE_ITEM ();                \
   DEBUG_PRINT2 ("  Popping high active reg: %d\n", high_reg);       \
                                     \
-  low_reg = (unsigned) POP_FAILURE_ITEM ();             \
+  low_reg = (unsigned) (size_t) POP_FAILURE_ITEM ();             \
   DEBUG_PRINT2 ("  Popping  low active reg: %d\n", low_reg);        \
                                     \
   for (this_reg = high_reg; this_reg >= low_reg; this_reg--)        \
@@ -3721,7 +3721,7 @@ int re_match_2 (struct re_pattern_buffer * bufp, const char * string1, int size1
                           regstart[r] = old_regstart[r];
 
                           /* xx why this test?  */
-                          if ((int) old_regend[r] >= (int) regstart[r])
+                          if (old_regend[r] >= regstart[r])
                             regend[r] = old_regend[r];
                         }
                     }
